test: add snapshot testing for cloud-rad tool
This changes adds initial snapshots based on the nodejs-deploy v1.0.0 library. These initial snapshots are derived from the existing docfx yaml from the staging bucket.

As part of this change, we have removed manually installing api-documenter, opting to install it through npm.

This change also modifies the generate script to not upload to docuploader if the NO_UPLOAD envvar is set.