refactor: rewrite YAML-generation pipeline
This PR is based on discussions with @tbpg and @dansaadati. It reflects the following goals:

+ Structure the code more clearly, so that it's easier to add features and fix bugs.
+ Use JavaScript, not shell.
+ During post-processing, treat YAML files as objects, rather than processing them line-by-line.
+ During post-processing, read, process, and write each YAML file only once.
+ Provide full test coverage.

My intent for this PR was to avoid making any meaningful changes to the YAML output. I think I've achieved that goal, based on eyeballing some diffs between the previous version's output and this version's output. Let me know if I should do more.